wip - split test stream, timeout in citgm
